# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Instantly Service",
    "description": "Cold email outreach service via Instantly.ai API. Manages campaigns, leads, accounts, analytics, and webhook events.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3011"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UpdateStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "completed"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "type": "object",
            "properties": {
              "emailsContacted": {
                "type": "number",
                "description": "Leads added to a campaign (row exists in instantly_campaigns, immediate)"
              },
              "emailsSent": {
                "type": "number",
                "description": "Total email_sent events (confirmed by Instantly webhook)"
              },
              "emailsDelivered": {
                "type": "number",
                "description": "emailsSent minus emailsBounced"
              },
              "emailsOpened": {
                "type": "number",
                "description": "Unique recipients who opened (COUNT DISTINCT lead_email with email_opened events)"
              },
              "emailsClicked": {
                "type": "number",
                "description": "Total link click events"
              },
              "emailsReplied": {
                "type": "number",
                "description": "Positive replies only (lead_interested events)"
              },
              "emailsBounced": {
                "type": "number",
                "description": "Total email_bounced events"
              },
              "repliesAutoReply": {
                "type": "number",
                "description": "Total auto_reply_received events"
              },
              "repliesNotInterested": {
                "type": "number",
                "description": "Total lead_not_interested events"
              },
              "repliesOutOfOffice": {
                "type": "number",
                "description": "Total lead_out_of_office events"
              },
              "repliesUnsubscribe": {
                "type": "number",
                "description": "Total lead_unsubscribed events"
              }
            },
            "required": [
              "emailsContacted",
              "emailsSent",
              "emailsDelivered",
              "emailsOpened",
              "emailsClicked",
              "emailsReplied",
              "emailsBounced",
              "repliesAutoReply",
              "repliesNotInterested",
              "repliesOutOfOffice",
              "repliesUnsubscribe"
            ]
          },
          "recipients": {
            "type": "number",
            "description": "Unique recipients (COUNT DISTINCT lead_email with email_sent events)"
          },
          "stepStats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "step": {
                  "type": "number",
                  "description": "Step number (1-based)"
                },
                "emailsSent": {
                  "type": "number"
                },
                "emailsOpened": {
                  "type": "number"
                },
                "emailsReplied": {
                  "type": "number"
                },
                "emailsBounced": {
                  "type": "number"
                }
              },
              "required": [
                "step",
                "emailsSent",
                "emailsOpened",
                "emailsReplied",
                "emailsBounced"
              ]
            },
            "description": "Per-step breakdown (only present when step data exists)"
          }
        },
        "required": [
          "stats",
          "recipients"
        ]
      },
...
      "StatusResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "leadId": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "campaign": {
                  "type": "object",
                  "nullable": true,
                  "properties": {
                    "lead": {
                      "type": "object",
                      "properties": {
                        "contacted": {
                          "type": "boolean"
                        },
                        "delivered": {
                          "type": "boolean"
                        },
                        "replied": {
                          "type": "boolean"
                        },
                        "lastDeliveredAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "contacted",
                        "delivered",
                        "replied",
                        "lastDeliveredAt"
                      ]
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "contacted": {
                          "type": "boolean"
                        },
                        "delivered": {
                          "type": "boolean"
                        },
                        "bounced": {
                          "type": "boolean"
                        },
                        "unsubscribed": {
                          "type": "boolean"
                        },
                        "lastDeliveredAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "contacted",
                        "delivered",
                        "bounced",
                        "unsubscribed",
                        "lastDeliveredAt"
                      ]
                    }
                  },
                  "required": [
                    "lead",
                    "email"
                  ]
                },
                "brand": {
                  "type": "object",
                  "properties": {
                    "lead": {
                      "type": "object",
                      "properties": {
                        "contacted": {
                          "type": "boolean"
                        },
                        "delivered": {
                          "type": "boolean"
                        },
                        "replied": {
                          "type": "boolean"
                        },
                        "lastDeliveredAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "contacted",
                        "delivered",
                        "replied",
                        "lastDeliveredAt"
                      ]
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "contacted": {
                          "type": "boolean"
                        },
                        "delivered": {
                          "type": "boolean"
                        },
                        "bounced": {
                          "type": "boolean"
                        },
                        "unsubscribed": {
                          "type": "boolean"
                        },
                        "lastDeliveredAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "contacted",
                        "delivered",
                        "bounced",
                        "unsubscribed",
                        "lastDeliveredAt"
                      ]
                    }
                  },
                  "required": [
                    "lead",
                    "email"
                  ]
                },
                "global": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "object",
                      "properties": {
                        "bounced": {
                          "type": "boolean"
                        },
                        "unsubscribed": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "bounced",
                        "unsubscribed"
                      ]
                    }
                  },
                  "required": [
                    "email"
                  ]
                }
              },
              "required": [
                "leadId",
                "email",
                "campaign",
                "brand",
                "global"
              ]
            }
          }
        },
        "required": [
          "results"
        ]
      },
      "StatusRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "description": "Brand ID to scope brand-level results (matches campaigns containing this brand)"
          },
          "campaignId": {
            "type": "string",
            "description": "Campaign ID to scope campaign-level results (omit for global-only checks)"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "leadId": {
                  "type": "string",
                  "description": "Lead-service lead ID (human)"
                },
                "email": {
                  "type": "string",
                  "description": "Email address"
                }
              },
              "required": [
                "leadId",
                "email"
              ]
            },
            "minItems": 1,
            "description": "Lead/email pairs to check"
          }
        },
        "required": [
          "brandId",
          "items"
        ]
      }
    },
...
  },
  "paths": {
...
    "/campaigns/{campaignId}/status": {
      "patch": {
        "summary": "Update campaign status",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "400": {
            "description": "Invalid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/stats": {
      "get": {
        "summary": "Get aggregated stats by filters",
        "description": "Aggregates stats from webhook events across campaigns matching the provided filters. Filters passed as query params; runIds is comma-separated.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of run IDs"
            },
            "required": false,
            "description": "Comma-separated list of run IDs",
            "name": "runIds",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by brand ID (matches campaigns containing this brand)"
            },
            "required": false,
            "description": "Filter by brand ID (matches campaigns containing this brand)",
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by exact workflow slug"
            },
            "required": false,
            "description": "Filter by exact workflow slug",
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by exact feature slug"
            },
            "required": false,
            "description": "Filter by exact feature slug",
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow dynasty slug (resolved to all versioned slugs)"
            },
            "required": false,
            "description": "Filter by workflow dynasty slug (resolved to all versioned slugs)",
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature dynasty slug (resolved to all versioned slugs)"
            },
            "required": false,
            "description": "Filter by feature dynasty slug (resolved to all versioned slugs)",
            "name": "featureDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "brandId",
                "campaignId",
                "workflowSlug",
                "featureSlug",
                "leadEmail",
                "workflowDynastySlug",
                "featureDynastySlug"
              ],
              "description": "Group results by dimension"
            },
            "required": false,
            "description": "Group results by dimension",
            "name": "groupBy",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated campaign stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "No filter provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/status": {
      "post": {
        "summary": "Batch delivery status check for leads/emails",
        "description": "Returns campaign-scoped, brand-scoped, and global delivery status for each lead/email pair. Campaign filters by campaignId (null if omitted); brand filters by brandId; global aggregates across everything.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delivery status results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}